requirements.txt
Pillow==10.0.0
requests==2.31.0
python-telegram-bot==20.6
yt-dlp>=2025.3.0
